Fixes to break-out of iterating libs.
